requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-google-genai
langchain-huggingface
langchain-pinecone
pinecone-client
pypdf
sentence-transformers
requests